FIX format and datatype

+++ tests/unit/ops/transformer/inference/test_bias_gelu.py
@@ -8,8 +8,7 @@ import deepspeed
 from deepspeed.ops.op_builder import UnitTestBuilder
 
 if not deepspeed.ops.__compatible_ops__[UnitTestBuilder.NAME]:
-    pytest.skip("Unittest ops are not available on this system",
-                allow_module_level=True)
+    pytest.skip("Unittest ops are not available on this system", allow_module_level=True)
 
 unittest_module = None
 torch_minor_version = None
@@ -47,11 +46,12 @@ def run_bias_gelu_ds(activations, bias):
     else:
         return unittest_module.bias_gelu_fp32(activations, bias)
 
+
 @pytest.mark.unittest
 @pytest.mark.parametrize("batch", [1, 2])
 @pytest.mark.parametrize("sequence", [1, 128, 255])
 @pytest.mark.parametrize("channels", [512, 1232, 4096])
-@pytest.mark.parametrize("dtype", [torch.float32, torch.float32])
+@pytest.mark.parametrize("dtype", [torch.float16, torch.float32])
 def test_bias_gelu(batch, sequence, channels, dtype):
     activations_ds = torch.randn((batch, sequence, channels), dtype=dtype, device='cuda')
     bias_ds = torch.randn((channels), dtype=dtype, device='cuda')
